creating the user types table

+++ src/schemas/items/items.service.ts
@@ -2,12 +2,16 @@ import { Inject, Injectable, forwardRef } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Item } from '../items/entities/item.entity';
 import { Repository } from 'typeorm';
-import { join } from 'path';
 import { CreateItemDto } from '../items/dto/create-item.dto';
 import { UpdateItemDto } from '../items/dto/update-item.dto';
 import { AuthorsService } from '../authors/authors.service';
 import { deleteFile, deleteFiles } from 'src/utils/storageProcess/deleteFiles';
 import { CategoriesService } from '../categories/categories.service';
+import {
+  filterNullsArray,
+  filterNullsObject,
+} from 'src/utils/helpers/filterNulls';
+import { mergeWithoutDups } from 'src/utils/helpers/mergeWithoutDups';
 
 @Injectable()
 export class ItemsService {
@@ -49,21 +53,6 @@ export class ItemsService {
     }
   }
 
-  downloadImage(imageName: string) {
-    try {
-      const response = join(process.cwd(), 'public/assets/items/' + imageName);
-      return {
-        message: response
-          ? 'Image returned successfully'
-          : "Item doesn't exist",
-        data: response,
-        status: response ? 200 : 404,
-      };
-    } catch (error) {
-      return { message: 'Error occurred', data: error, status: 500 };
-    }
-  }
-
   async createItem(createItemDto: CreateItemDto) {
     try {
       // check the author and category
@@ -84,15 +73,12 @@ export class ItemsService {
       // create the item
       const newItem = this.itemRepository.create({
         ...createItemDto,
-        image: createItemDto.image.filename || '',
-        screenshots: createItemDto.screenshots.map(
-          (screenshot) => screenshot.filename || '',
+        image: createItemDto?.image?.filename || '',
+        screenshots: createItemDto?.screenshots?.map(
+          (screenshot) => screenshot?.filename || '',
         ),
       });
       const response = await this.itemRepository.save(newItem);
-
-      // append the item to author
-      // this.authorsService.appendItem(newItem.id, newItem.screenshots);
 
       return {
         message: 'Item has been created successfully',
@@ -110,40 +96,50 @@ export class ItemsService {
 
   async updateItem(id: string, updateItemDto: UpdateItemDto) {
     try {
-      const item = await this.getItemById(updateItemDto.authorId);
+      const item = await this.getItemById(id);
       const author = await this.authorsService.getAuthorById(
         updateItemDto.authorId,
       );
+      const category = await this.categoriesService.getCategoryById(
+        updateItemDto.categoryId,
+      );
 
-      if (!item || !author) {
+      if (!item || !author || !category) {
         return {
           message: 'Invalid data',
-          data: `Provided ${!item ? 'item' : 'author'} does not exist`,
-          status: 400,
+          data: `Provided ${
+            !item ? 'item' : !author ? 'author' : 'category'
+          } does not exist`,
+          status: 404,
         };
+      }
+
+      const oldArray: string[] = item.data.screenshots;
+
+      const newArray = filterNullsArray(updateItemDto?.screenshots).map(
+        (screenshot: { filename: string }) => screenshot.filename,
+      );
+
+      const newObject = filterNullsObject({
+        ...updateItemDto,
+        image: updateItemDto.image.filename,
+      });
+
+      if (newArray.length > 0) {
+        newObject['screenshots'] = mergeWithoutDups([oldArray, newArray]);
       }
 
       const response = await this.itemRepository.update(
         {
           id,
         },
-        {
-          ...updateItemDto,
-          image: updateItemDto.image.filename || '',
-          screenshots: updateItemDto.screenshots.map(
-            (screenshot) => screenshot.filename || '',
-          ),
-        },
+        newObject,
       );
 
-      const isItemExist = response.affected !== 0;
-
       return {
-        message: isItemExist
-          ? 'Item has been updated successfully'
-          : "Item doesn't exist",
+        message: 'Item has been updated successfully',
         data: response,
-        status: isItemExist ? 200 : 404,
+        status: 200,
       };
     } catch (error) {
       return {
@@ -184,15 +180,15 @@ export class ItemsService {
         };
       }
 
-      const imageName: string = item?.data?.image;
       const screenshotsNames: string[] = item?.data?.screenshots;
       const response = await this.itemRepository.delete(id);
 
       // delete the images related to the file
-      deleteFile('./public/assets/items/' + imageName);
+      item?.data?.image &&
+        deleteFile('./public/assets/items/' + item?.data?.image);
 
       screenshotsNames.forEach((screenshot) => {
-        deleteFile('./public/assets/items/' + screenshot);
+        screenshot && deleteFile('./public/assets/items/' + screenshot);
       });
 
       return {
