+++ src/schemas/users/users.service.ts
@@ -1,4 +1,3 @@
-/* eslint-disable @typescript-eslint/no-unused-vars */
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
@@ -6,15 +5,38 @@ import { User } from './entities/user.entity';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { deleteFile, deleteFiles } from 'src/utils/storageProcess/deleteFiles';
-import { join } from 'path';
-import { hash } from 'bcrypt';
+import { compare, hash } from 'bcrypt';
+import { filterNullsObject } from 'src/utils/helpers/filterNulls';
+import { JwtService } from '@nestjs/jwt';
+import { FullTokenPayload, TokenPayload } from 'src/types/token-payload.type';
+import { Request } from 'express';
+import { UserRole } from 'src/enums/user-role.enum';
 
 @Injectable()
 export class UsersService {
   constructor(
     @InjectRepository(User)
     private readonly userRepository: Repository<User>,
+    private jwtService: JwtService,
   ) {}
+
+  private passwordRemover(user: User) {
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    const { password, ...rest } = user;
+    return rest;
+  }
+
+  getUserTokenData(req: Request) {
+    const userTokenData = {
+      ...req?.user,
+    } as FullTokenPayload;
+
+    userTokenData['expiredIn'] = `${Math.floor(
+      (userTokenData.exp - userTokenData.iat) / 3600,
+    )} Hours`;
+
+    return userTokenData;
+  }
 
   async getUsers(conditions: Record<string, any>, withPass: boolean = false) {
     try {
@@ -22,8 +44,7 @@ export class UsersService {
 
       // remove the password from all the users before sending response
       const updatedUsers = response.map((user) => {
-        const { password, ...rest } = user;
-        return withPass ? user : rest;
+        return withPass ? user : this.passwordRemover(user);
       });
 
       return {
@@ -50,12 +71,9 @@ export class UsersService {
         };
       }
 
-      // remove the password from the user before sending response
-      const { password, ...deprecatedPassUser } = response;
-
       return {
         message: 'User has been found',
-        data: deprecatedPassUser,
+        data: this.passwordRemover(response),
         status: 200,
       };
     } catch (error) {
@@ -63,23 +81,23 @@ export class UsersService {
     }
   }
 
-  downloadImage(imageName: string) {
+  async createUser(
+    createUserDto: CreateUserDto,
+    userTokenData: FullTokenPayload,
+  ) {
     try {
-      const response = join(process.cwd(), 'public/assets/users/' + imageName);
-      return {
-        message: response
-          ? 'Image returned successfully'
-          : "User doesn't exist",
-        data: response,
-        status: response ? 200 : 404,
-      };
-    } catch (error) {
-      return { message: 'Error occurred', data: error, status: 500 };
-    }
-  }
+      if (
+        createUserDto.role === UserRole.ADMIN &&
+        (!userTokenData || userTokenData.role === UserRole.MEMBER)
+      ) {
+        return {
+          message:
+            'Unauthorized entrance, you must be an admin to create another admin account',
+          data: { token: userTokenData },
+          status: 401,
+        };
+      }
 
-  async createUser(createUserDto: CreateUserDto) {
-    try {
       const hashedPass = await hash(createUserDto.password, 12);
 
       const newUser = this.userRepository.create({
@@ -91,7 +109,7 @@ export class UsersService {
 
       return {
         message: 'User has been created successfully',
-        data: response,
+        data: this.passwordRemover(response),
         status: 200,
       };
     } catch (error) {
@@ -103,25 +121,43 @@ export class UsersService {
     }
   }
 
-  async updateUser(id: string, updateUserDto: UpdateUserDto) {
+  async updateUser(
+    id: string,
+    updateUserDto: UpdateUserDto,
+    userTokenData: FullTokenPayload,
+  ) {
     try {
       const user = await this.getUserById(id);
       if (!user) {
         return {
           message: 'Invalid data',
           data: `Provided user does not exist`,
-          status: 400,
+          status: 404,
         };
       }
+
+      if (
+        userTokenData.userId !== id &&
+        userTokenData.role !== UserRole.ADMIN
+      ) {
+        return {
+          message:
+            "Unauthorized entrance, you're only allowed to update your account",
+          data: { token: userTokenData, id },
+          status: 401,
+        };
+      }
+
+      const newObject = filterNullsObject({
+        ...updateUserDto,
+        avatar: updateUserDto?.avatar?.filename,
+      });
 
       const response = await this.userRepository.update(
         {
           id,
         },
-        {
-          ...updateUserDto,
-          avatar: updateUserDto.avatar?.filename || '',
-        },
+        newObject,
       );
 
       return {
@@ -157,7 +193,7 @@ export class UsersService {
     }
   }
 
-  async deleteUser(id: string) {
+  async deleteUser(id: string, userTokenData: FullTokenPayload) {
     try {
       const user = await this.getUserById(id);
       if (user.status === 404) {
@@ -168,10 +204,23 @@ export class UsersService {
         };
       }
 
+      if (
+        userTokenData.userId !== id &&
+        userTokenData.role !== UserRole.ADMIN
+      ) {
+        return {
+          message:
+            "Unauthorized entrance, you're only allowed to delete your account",
+          data: { token: userTokenData, id },
+          status: 401,
+        };
+      }
+
       const response = await this.userRepository.delete(id);
 
       // delete the image related to the file
-      deleteFile('./public/assets/users/' + user?.data?.avatar);
+      user?.data?.avatar &&
+        deleteFile('./public/assets/users/' + user?.data?.avatar);
 
       return {
         message: 'User has been deleted successfully',
@@ -180,6 +229,47 @@ export class UsersService {
       };
     } catch (error) {
       return { message: 'Error occurred', data: error, status: 500 };
+    }
+  }
+
+  async logIn(email: string, password: string) {
+    try {
+      const response = await this.getUsers({ email }, true);
+      if (response?.data?.length === 0) {
+        return {
+          message: 'Invalid email',
+          data: email,
+          status: 400,
+        };
+      }
+
+      const user: User = response.data[0];
+      if (!(await compare(password, user?.password))) {
+        return {
+          message: 'Invalid password',
+          data: password,
+          status: 400,
+        };
+      }
+
+      const payload: TokenPayload = {
+        userId: user?.id,
+        username: user?.userName,
+        role: user?.role,
+        email: user?.email,
+      };
+
+      return {
+        message: 'Token has been generated',
+        data: await this.jwtService.signAsync(payload),
+        status: 200,
+      };
+    } catch (error) {
+      return {
+        message: 'Error occurred',
+        data: error,
+        status: 500,
+      };
     }
   }
 }
